fix: invoke matching workflow as a job
